# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api_handlers.createTrackRequest:
    properties:
      album_id:
        type: string
      artist_id:
        type: string
      bitrate:
        type: integer
      disc_number:
        type: integer
      duration_ms:
        type: integer
      file_size:
        type: integer
      format:
        type: string
      title:
        type: string
      track_number:
        type: integer
    type: object
...
  internal_api_handlers.trackResponse:
    properties:
      album_id:
        type: string
      artist_id:
        type: string
      bitrate:
        type: integer
      created_at:
        type: string
      disc_number:
        type: integer
      duration_ms:
        type: integer
      file_size:
        type: integer
      format:
        type: string
      has_file:
        description: |-
          HasFile says whether an upload has landed, without exposing
          storage_key, which names an object in the bucket - layout, not API.
        type: boolean
      id:
        type: string
      title:
        type: string
      track_number:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Music library management and streaming API
  title: Mydal API
  version: "1.0"
paths:
...
  /tracks:
    post:
      consumes:
      - application/json
      description: Add a new track to the database
      parameters:
      - description: Track payload
        in: body
        name: track
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.createTrackRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handlers.trackResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new track
      tags:
      - tracks
  /tracks/{id}:
    delete:
      description: Remove a track from the database by ID
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a track
      tags:
      - tracks
    get:
      description: Retrieve a single track by its unique identifier
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.trackResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get track by ID
      tags:
      - tracks
  /tracks/{id}/file:
    put:
      consumes:
      - application/octet-stream
      description: Upload the audio binary for an existing track. The format is detected
        from the file's own bytes; the Content-Type header is ignored. A chunked body
        without Content-Length is accepted.
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: string
      - description: Raw audio bytes, sent as the request body - not a multipart form
        in: body
        name: audio
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload track audio file
      tags:
      - tracks
  /tracks/{id}/stream:
    get:
      description: Stream the audio file for a track via HTTP range requests
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - audio/*
      responses:
        "200":
          description: OK
          schema:
            type: file
        "206":
          description: Partial Content
          schema:
            type: file
        "304":
          description: Not Modified
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "416":
          description: 'Plain text, not JSON: written by net/http''s own range handling'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream track audio
      tags:
      - streaming
swagger: "2.0"